Refactor: 일정 생성 트랜잭션 경계 분리
일정 생성의 데이터 조회, consensus/LLM 계산, 저장 단계를 분리해 외부 API 및 장시간 계산이 DB 쓰기 트랜잭션을 점유하지 않도록 조정.

Constraint: 일정 생성 결과 API 응답 구조 유지

Rejected: 기존 ScheduleService 단일 @Transactional 유지 | LLM refinement 가능 구간이 DB 커넥션 점유

Confidence: high

Scope-risk: moderate

Tested: ./gradlew test --rerun-tasks --no-daemon

Not-tested: 실제 OpenAI API 호출 성공 경로
